move configtest to integration test
